src/requirements.txt
Flask==3.0.0
google-api-python-client==2.115.0
gunicorn==22.0.0
google-cloud-storage==2.16.0
